Memcache.Client

Binary protocol client for Memcached server.

Types

key :: binary

opts :: Keyword.t

value :: any

Functions

add(key, value, opts \\ [])

Specs

add(key, value, opts) :: Memcache.Client.Response.t

Sets value for given key only if it does not already exist.

append(key, value)

Specs

append(key, value) :: Memcache.Client.Response.t

Appends value to given key if it already exists.

decrement(key, amount, opts \\ [])

Specs

decrement(key, pos_integer, opts) :: Memcache.Client.Response.t

Decrements a counter on given key.

delete(key)

Specs

delete(key) :: Memcache.Client.Response.t

Deletes the value for the given key.

flush(opts \\ [])

Specs

flush(opts) :: Memcache.Client.Response.t

Flushes the cache.

get(key)

Specs

get(key) :: Memcache.Client.Response.t

Gets value for given key.

increment(key, amount, opts \\ [])

Specs

increment(key, pos_integer, opts) :: Memcache.Client.Response.t

Increments a counter on given key.

mget(keys)

Specs

mget(Enumerable.t) :: Stream.t

Gets values for multiple keys with a single pipelined operation.

mset(keyvalues)

Specs

mset(Enumerable.t) :: Stream.t

Sets multiple values with a single pipelined operation. Value needs to be a tuple of key and value.

prepend(key, value)

Specs

prepend(key, value) :: Memcache.Client.Response.t

Prepends value to given key if it already exists.

replace(key, value, opts \\ [])

Specs

replace(key, value, opts) :: Memcache.Client.Response.t

Sets valuefor given key only if it already exists.

set(key, value, opts \\ [])

Specs

set(key, value, opts) :: Memcache.Client.Response.t

Sets value for given key.

start(type, args)

Callback implementation for c::application.start/2.

version()

Specs

version :: Memcache.Client.Response.t

Returns the current memcached version.